+++ udata/core/dataset/api_fields.py
@@ -284,6 +284,7 @@ DEFAULT_MASK = ",".join(
         "quality",
         "internal",
         "contact_points",
+        "featured",
     )
 )